requirements.txt
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
httplib2==0.22.0
idna==3.10
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
six==1.17.0
tornado==6.4.2
urllib3==2.2.3
watchdog==6.0.0
